Reduced minSdkVersion significantly.
Whoops.  Didn't even notice it defaulted to 23.  To that end, the compat
library for VectorDrawable is now in play for anything earlier than 21.

As it stands, I think the code annotations mean I can't go below 14 for
the time being, but 14 should support a lot more stuff.  If I can get to
7 (the bare minimum for Live Wallpapers), that'd be even better.