Added newer Omega Shift banner.
Apparently for 2021 they updated the Omega Shift banner, as per the
Pinny Arcade pin in the merch store and a vector image in the media kit.
This adds that banner to the wallpaper.

This also adds an option to keep using the old quartered-circle-on-black
version, if the user so chooses.